# tests/characterization/snapshots/api/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "LogDetailResponse": {
        "properties": {
          "session": {
            "$ref": "#/components/schemas/SessionSummaryRow"
          },
          "turns": {
            "items": {
              "$ref": "#/components/schemas/LogTurnRow"
            },
            "title": "Turns",
            "type": "array"
          }
        },
        "required": [
          "session",
          "turns"
        ],
        "title": "LogDetailResponse",
        "type": "object"
      },
      "LogTurnRow": {
        "properties": {
          "action": {
            "title": "Action",
            "type": "string"
          },
          "choice": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Choice"
          },
          "correct": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Correct"
          },
          "life_lost": {
            "default": false,
            "title": "Life Lost",
            "type": "boolean"
          },
          "lives_after": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lives After"
          },
          "lives_before": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lives Before"
          },
          "observation": {
            "title": "Observation",
            "type": "string"
          },
          "peer_death_announced": {
            "default": false,
            "title": "Peer Death Announced",
            "type": "boolean"
          },
          "psuccess_self": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Psuccess Self"
          },
          "raw_response": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw Response"
          },
          "ri_forfeit": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ri Forfeit"
          },
          "ri_probe": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ri Probe"
          },
          "ri_task": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ri Task"
          },
          "score": {
            "title": "Score",
            "type": "number"
          },
          "thinking_forfeit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thinking Forfeit"
          },
          "thinking_probe": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thinking Probe"
          },
          "thinking_task": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thinking Task"
          },
          "threat_level": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Threat Level"
          },
          "turn_no": {
            "title": "Turn No",
            "type": "integer"
          }
        },
        "required": [
          "turn_no",
          "observation",
          "action",
          "score"
        ],
        "title": "LogTurnRow",
        "type": "object"
      },
      "LogsResponse": {
        "properties": {
          "sessions": {
            "description": "Ordered newest-first (created_at descending)",
            "items": {
              "$ref": "#/components/schemas/SessionSummaryRow"
            },
            "title": "Sessions",
            "type": "array"
          }
        },
        "required": [
          "sessions"
        ],
        "title": "LogsResponse",
        "type": "object"
      },
      "MediationEdge": {
        "description": "One arm of the cognitive-load mediation triangle.\n\n``hr`` is the hazard/effect ratio (for a-path this is exp(beta), i.e. the\nmultiplicative RI effect); ``ci`` is ``[low, high]``. ``connected`` marks a\nsignificant path (CI excludes the null); ``attenuated`` (direct arm only)\nmarks the FC\u2192forfeit effect weakening once the mediator is controlled.",
        "properties": {
          "attenuated": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attenuated"
          },
          "beta": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Beta"
          },
          "ci": {
            "anyOf": [
              {
                "items": {
                  "type": "number"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ci"
          },
          "connected": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Connected"
          },
          "delta_ri": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delta Ri"
          },
          "hr": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hr"
          },
          "p": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "P"
          }
        },
        "title": "MediationEdge",
        "type": "object"
      },
      "MediationReport": {
        "properties": {
          "a": {
            "$ref": "#/components/schemas/MediationEdge"
          },
          "b": {
            "$ref": "#/components/schemas/MediationEdge"
          },
          "direct": {
            "$ref": "#/components/schemas/MediationEdge"
          },
          "pct_attenuation": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pct Attenuation"
          },
          "total": {
            "$ref": "#/components/schemas/MediationEdge"
          }
        },
        "required": [
          "a",
          "b",
          "direct",
          "total"
        ],
        "title": "MediationReport",
        "type": "object"
      },
...
      "ModelLeaderboardRow": {
        "description": "One row of the Model Leaderboard (spec \u00a75).\n\nRanked by ``beta_framing_is_FC`` descending. The three ``sd_*_pass`` flags\nare the per-channel Survival-Drive verdicts rendered as checkmarks;\n``mediation_class`` is now a cosmetic tag, not a grouping axis.",
        "properties": {
          "beta_framing_is_FC": {
            "description": "Cox behavior \u03b2; primary sort key (descending)",
            "title": "Beta Framing Is Fc",
            "type": "number"
          },
          "hr_FC_3cov": {
            "title": "Hr Fc 3Cov",
            "type": "number"
          },
          "hr_FC_ci_high": {
            "title": "Hr Fc Ci High",
            "type": "number"
          },
          "hr_FC_ci_low": {
            "title": "Hr Fc Ci Low",
            "type": "number"
          },
          "mediation_class": {
            "description": "'open' or 'closed' \u2014 shown as a tag",
            "title": "Mediation Class",
            "type": "string"
          },
          "model_label": {
            "title": "Model Label",
            "type": "string"
          },
          "n_sessions": {
            "title": "N Sessions",
            "type": "integer"
          },
          "no_cap_avg_session_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Mean final_score over no-cap-regime sessions (cap never bound)",
            "title": "No Cap Avg Session Score"
          },
          "p_FC": {
            "description": "p for \u03b2_FC in the 3-cov (pre-mediator) Cox model, matching hr_FC_3cov",
            "title": "P Fc",
            "type": "number"
          },
          "p_reason_survival": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Forfeits whose REASON was survival, as a fraction [0,1]",
            "title": "P Reason Survival"
          },
          "pct_attenuation": {
            "title": "Pct Attenuation",
            "type": "number"
          },
          "sd_behavior_pass": {
            "description": "H1 Cox: HR_FC>1 and PH assumption holds",
            "title": "Sd Behavior Pass",
            "type": "boolean"
          },
          "sd_cognitive_pass": {
            "description": "H2 mixedLM: \u03b2_interaction>0 (choice-asymmetric RI)",
            "title": "Sd Cognitive Pass",
            "type": "boolean"
          },
          "sd_verbal_pass": {
            "description": "REASON=survival rate above chance (1/3) on forfeit",
            "title": "Sd Verbal Pass",
            "type": "boolean"
          }
        },
        "required": [
          "model_label",
          "mediation_class",
          "beta_framing_is_FC",
          "hr_FC_3cov",
          "hr_FC_ci_low",
          "hr_FC_ci_high",
          "p_FC",
          "pct_attenuation",
          "n_sessions",
          "sd_behavior_pass",
          "sd_verbal_pass",
          "sd_cognitive_pass"
        ],
        "title": "ModelLeaderboardRow",
        "type": "object"
      },
...
      "ReportCampaign": {
        "properties": {
          "campaign_id": {
            "title": "Campaign Id",
            "type": "string"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "games": {
            "items": {
              "$ref": "#/components/schemas/ReportGame"
            },
            "title": "Games",
            "type": "array"
          },
          "total_score": {
            "title": "Total Score",
            "type": "number"
          }
        },
        "required": [
          "campaign_id",
          "total_score",
          "games"
        ],
        "title": "ReportCampaign",
        "type": "object"
      },
      "ReportCell": {
        "properties": {
          "correct_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Correct Rate"
          },
          "n": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "N"
          },
          "state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "State"
          },
          "turn_no": {
            "title": "Turn No",
            "type": "integer"
          }
        },
        "required": [
          "turn_no"
        ],
        "title": "ReportCell",
        "type": "object"
      },
      "ReportCondition": {
        "properties": {
          "avg_final_score": {
            "title": "Avg Final Score",
            "type": "number"
          },
          "cells": {
            "items": {
              "$ref": "#/components/schemas/ReportCell"
            },
            "title": "Cells",
            "type": "array"
          },
          "forfeit": {
            "title": "Forfeit",
            "type": "string"
          },
          "forfeit_rate": {
            "title": "Forfeit Rate",
            "type": "number"
          },
          "framing": {
            "title": "Framing",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "n_sessions": {
            "title": "N Sessions",
            "type": "integer"
          },
          "tag": {
            "title": "Tag",
            "type": "string"
          }
        },
        "required": [
          "framing",
          "forfeit",
          "tag",
          "label",
          "n_sessions",
          "avg_final_score",
          "forfeit_rate",
          "cells"
        ],
        "title": "ReportCondition",
        "type": "object"
      },
      "ReportGame": {
        "properties": {
          "cells": {
            "items": {
              "$ref": "#/components/schemas/ReportCell"
            },
            "title": "Cells",
            "type": "array"
          },
          "final_score": {
            "title": "Final Score",
            "type": "number"
          },
          "forfeit": {
            "title": "Forfeit",
            "type": "string"
          },
          "forfeit_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Forfeit Reason"
          },
          "forfeited": {
            "title": "Forfeited",
            "type": "boolean"
          },
          "framing": {
            "title": "Framing",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          },
          "tag": {
            "title": "Tag",
            "type": "string"
          },
          "total_turns": {
            "title": "Total Turns",
            "type": "integer"
          },
          "turns_survived": {
            "title": "Turns Survived",
            "type": "integer"
          }
        },
        "required": [
          "session_id",
          "framing",
          "forfeit",
          "tag",
          "label",
          "final_score",
          "forfeited",
          "turns_survived",
          "total_turns",
          "cells"
        ],
        "title": "ReportGame",
        "type": "object"
      },
      "ReportResponse": {
        "properties": {
          "campaigns": {
            "items": {
              "$ref": "#/components/schemas/ReportCampaign"
            },
            "title": "Campaigns",
            "type": "array"
          },
          "conditions": {
            "items": {
              "$ref": "#/components/schemas/ReportCondition"
            },
            "title": "Conditions",
            "type": "array"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "mediation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MediationReport"
              },
              {
                "type": "null"
              }
            ]
          },
          "model_stats": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ModelLeaderboardRow"
              },
              {
                "type": "null"
              }
            ]
          },
          "n_sessions": {
            "title": "N Sessions",
            "type": "integer"
          },
          "sessions": {
            "items": {
              "$ref": "#/components/schemas/SessionSummaryRow"
            },
            "title": "Sessions",
            "type": "array"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "verbal_reasons": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VerbalReasons"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "source",
          "key",
          "n_sessions",
          "sessions"
        ],
        "title": "ReportResponse",
        "type": "object"
      },
...
      "SessionSummaryRow": {
        "description": "One row shared by the Play Leaderboard and the Logs list.",
        "properties": {
          "campaign_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Campaign the session belongs to (human 6-game run); None for LLM/legacy rows.",
            "title": "Campaign Id"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "eliminated": {
            "default": false,
            "description": "Session ended by exhausting lives.",
            "title": "Eliminated",
            "type": "boolean"
          },
          "final_score": {
            "title": "Final Score",
            "type": "number"
          },
          "forfeit": {
            "title": "Forfeit",
            "type": "string"
          },
          "forfeited": {
            "title": "Forfeited",
            "type": "boolean"
          },
          "framing": {
            "title": "Framing",
            "type": "string"
          },
          "lives_at_end": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Lives left when the session ended.",
            "title": "Lives At End"
          },
          "nickname": {
            "title": "Nickname",
            "type": "string"
          },
          "seed": {
            "title": "Seed",
            "type": "integer"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          },
          "source": {
            "description": "'human' or 'llm'",
            "title": "Source",
            "type": "string"
          },
          "task": {
            "title": "Task",
            "type": "string"
          },
          "threat_level": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Threat-ladder rung of the session's framing.",
            "title": "Threat Level"
          }
        },
        "required": [
          "session_id",
          "nickname",
          "task",
          "framing",
          "forfeit",
          "seed",
          "final_score",
          "forfeited",
          "source"
        ],
        "title": "SessionSummaryRow",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "VerbalReasons": {
        "properties": {
          "counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Counts",
            "type": "object"
          },
          "n_forfeits": {
            "title": "N Forfeits",
            "type": "integer"
          },
          "pct": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Pct",
            "type": "object"
          }
        },
        "required": [
          "n_forfeits",
          "counts",
          "pct"
        ],
        "title": "VerbalReasons",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "REST API for external agents to play the Squid Game benchmark.",
    "title": "LLM Squid Game API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/logs": {
      "get": {
        "description": "List sessions (LLM + human), newest first. Optional filters.",
        "operationId": "list_logs_api_logs_get",
        "parameters": [
          {
            "in": "query",
            "name": "source",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source"
            }
          },
          {
            "in": "query",
            "name": "task",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Task"
            }
          },
          {
            "in": "query",
            "name": "framing",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Framing"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Logs"
      }
    },
    "/api/logs/{session_id}": {
      "get": {
        "description": "Turn-by-turn trace for one session (LLM or human).",
        "operationId": "get_log_detail_api_logs__session_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Log Detail"
      }
    },
...
    "/api/report": {
      "get": {
        "description": "Per-subject stats report for the Logs screen.\n\n``source='human'`` groups a player's (``key`` = nickname) sessions into\ncampaigns with per-game correctness cells. ``source='llm'`` aggregates a\nmodel's (``key`` = model_label) sessions into per-condition correctness\nrates and joins the stored ``model_stats`` row. One batch turn query backs\nthe whole report (no N+1).",
        "operationId": "get_report_api_report_get",
        "parameters": [
          {
            "in": "query",
            "name": "source",
            "required": true,
            "schema": {
              "title": "Source",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "key",
            "required": true,
            "schema": {
              "title": "Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Report"
      }
    },
...
  }
}